backend/requirements.txt
# backend/requirements.txt
# Metomi 后端依赖（锁定版本）

# ── Web 框架 ──
fastapi==0.131.0
uvicorn[standard]==0.41.0

# ── ORM & 数据库 ──
sqlalchemy[asyncio]==2.0.46
aiosqlite==0.22.1
alembic==1.18.4

# ── 速率限制 ──
slowapi==0.1.9

# ── 鉴权与安全 ──
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.5.0
python-multipart==0.0.22
bcrypt==4.0.1

# ── 文件处理与元数据 ──
ebooklib==0.20
beautifulsoup4==4.14.3
httpx==0.28.1
lxml==6.0.2

# ── 数据校验与配置 ──
pydantic==2.12.5
pydantic-settings==2.13.1
python-dotenv==1.2.1
PyMuPDF==1.25.3

# ── 文件异步 I/O ──
aiofiles==24.1.0

# ── 测试 ──
pytest==9.0.2
pytest-asyncio==1.3.0
